#31820 ドラム演奏画面のFullHD化を一通り完了した。以下の制限有り。 ・ライフゲージに横線が入る ・Risky時の「残りMiss許容回数」の数字の表示位置・表示サイズがおかしい ・グラフの表示未確認 ・スクロールの滑らかさが不足している(VGAでの表示座標をそのまま定数倍しているだけだからこうなります) #31820 EXIT画面をFullHD化した。(jpegファイルを1枚差し替えただけ)
git-svn-id: http://svn.osdn.jp/svnroot/dtxmania/branches/130731(DTXMania097%20with%2023928%20FullHD)@569 16f42ceb-6dc6-49c8-ba94-f2d53467949d

+++ DTXManiaプロジェクト/コード/ステージ/07.演奏/CAct演奏AVI.cs
@@ -278,13 +278,12 @@ namespace DTXMania
 					CDTXMania.app.GraphicsDeviceManager.CurrentSettings.BackBufferFormat, Pool.Default, Usage.Dynamic );
 #else
 				this.tx描画用 = new CTexture( CDTXMania.app.Device,
-					//(int) ( 278 * Scale.X ),
-					//(int) ( 355 * Scale.Y ),
-					(int) ( 278 ),
-					(int) ( 355 ),
+					278,
+					355,
 					CDTXMania.app.GraphicsDeviceManager.CurrentSettings.BackBufferFormat, Pool.Managed );
 #endif
 				this.tx描画用.vc拡大縮小倍率 = new Vector3( Scale.X, Scale.Y, 1f );
+				//this.tx描画用.vc拡大縮小倍率 = new Vector3( 2f, 2f, 1f );
 				base.OnManagedリソースの作成();
 			}
 		}
